==> 03_cox model.Rmd <==
---
title: "03_cox model"
author: "Brenna Kelly"
date: "2025-06-03"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}


library(dplyr)
library(purrr)
library(regclass)
library(data.table)

# OG
full_dat <- read.csv("data/clean_birth_data.csv")
# outcome: placental insufficiency
full_dat$hypertension_code <- ifelse(full_dat$hypertension %in% c("C", "E", "H", "P"),
                                     "hyp", "no")
full_dat$pl_insuf <- case_when(full_dat$sga10_by_sex > full_dat$birthweightgrams ~ 1, # sga
                               full_dat$hypertension %in% c("H", "E", "P") ~ 1)
full_dat$pl_insuf <- ifelse(is.na(full_dat$pl_insuf), 0, 1)

# outcome: gestational hypertension
full_dat$gestational_hyp <- ifelse(full_dat$hypertension %in% c("H", "E", "P"), 1, 0)

# outcome: gestational diabetes
full_dat$gestational_diab <- ifelse(full_dat$gestationaldiabetes == 1, 1, 0)

# outcome: preterm birth
full_dat$preterm <- ifelse(full_dat$gestation < 34, "preterm", "term")
full_dat$preterm_code <- ifelse(full_dat$preterm == "preterm", 1, 0)

# outcome: very preterm birth
full_dat$very_preterm <- ifelse(full_dat$gestation < 30, "preterm", "term")

# *negative control* outcome: gender
full_dat$gender_code <- ifelse(full_dat$gender == "M", 1, 0)

# grab confounders
dat <- fread("/Users/brenna/Downloads/birth_exposures_with_confounders.csv")

confounders <- dat[, c("momid", "grid_id", "gestation", "birthccyy", "birthmm", 
                       "mommedicaid_char", "mommaritalstatus_char",
                       "momwic_char", "momeducation_char", "momage",
                       "poverty_rate")]
rm(dat)

# and a few more
# mom <- read.csv('../../data/health data/births.csv')
mom_2013 <- fread('/Users/brenna/Downloads/births 2013+.csv')
mom_2016 <- fread('/Users/brenna/Downloads/births 2016+.csv')

mom <- rbind(mom_2013, mom_2016)

mom <- mom |>
  mutate(MomHeightFeet = as.numeric(MomHeightFeet),
         MomHeightFeet = ifelse(MomHeightFeet == 1, NA, MomHeightFeet),
         MomHeightInches = as.numeric(MomHeightInches),
         race_ethn = case_when(MomHispanicOrigin == 1 | MomHispanicMex == 1 |
                                 MomHispanicCuban == 1 | MomHispanicPuertoRican == 1 |
                                 MomHispanicOther == 1 | MomHispanicOtherSpec != "" |
                                 !MomHispanicOtherSpec %in% c("*", "") ~ # primacy
                                 "hispanic",
                               MomRaceWhite == 1 ~ "white",
                               MomRaceBlack == 1 ~ "black",
                               MomRaceAmIndian == 1 ~ "amindian",
                               MomRaceChinese == 1 | MomRaceJapanese == 1 |
                                 MomRaceFilipino == 1 | MomRaceOtherAsian == 1 |
                                 MomRaceAsianIndian == 1 | MomRaceKorean == 1 |
                                 MomRaceVietnamese == 1 ~ "asian",
                               MomRaceHawaiian == 1 | MomRaceSamoan == 1 |
                                 MomRaceGuamanian == 1 | MomRacePacIslander == 1 |
                                 MomRaceTongan == 1 ~ "nhopi",
                               MomRaceOther == 1 | MomRaceOtherSpec1 == 1 |
                                 MomRaceOtherSpec2 == 1 | MomRaceUnknown == 1 ~ "other"),
         race_ethn = ifelse(is.na(race_ethn), "other", race_ethn),
         # there's one maternal_wt = 550; 
         # this could be a recording error, but it's prior to 2013 so don't bother with it
         MomWeightPrior = as.numeric(MomWeightPrior),
         # bmi = ((lbs * 703) / inches) / inches
         mom_bmi = ((MomWeightPrior * 703) / 
                      (12 * MomHeightFeet + MomHeightInches)) / 
           (12 * MomHeightFeet + MomHeightInches))

names(mom) <- tolower(names(mom))
mom <- mom[, c("momid", "birthccyy", "birthmm",
               "race_ethn")] #, "mom_bmi", "momheightfeet",
               #"momheightinches", "momweightprior"
               # note: too much missing in BMI
mom$momid <- as.numeric(mom$momid)

# and a couple more
table(full_dat[, c("prevpregnowliving", "prevpregnowdead", "prevpregstillborn")])
# none — these were excluded; our cohort is nulligravida (as far as we know)

# birth data
full_dat <- merge(full_dat, mom, by = c("momid", "birthccyy", "birthmm"))
z <- read.csv("data/birth_traj_trunc_jun08.csv")

z_y <- merge(z, full_dat[, c("momid", "grid_id", "birthccyy", "birthmm",
                             "sga", "hypertension_code",
                             "pl_insuf", "preterm", "preterm_code",
                             "gestational_hyp", "very_preterm",
                             "gestational_diab", "gender_code",
                             "race_ethn")], 
             by = c("momid", "grid_id")) # N = 44984

dat_con <- merge(z_y, confounders, by = c("momid", "grid_id", 
                                          "birthccyy", "birthmm"))

dat_con$poverty_rate <- ifelse(is.na(dat_con$poverty_rate), 0, dat_con$poverty_rate)
dat_con$poverty_rate_sc <- (dat_con$poverty_rate - mean(dat_con$poverty_rate)) / 10

# drop subsequent pregnancies
test <- dat_con |>
  arrange(momid, birthccyy, birthmm)

dat_con <- test[match(unique(test$momid), test$momid), ]

## multiples, for sanity check
# moms_multiple
## 518477548 should be 10/2014
## 519421022 should be 11/2013
## 529739095 should be 4/2013
## 529897525 should be 8/2015
# t.first |>
#   filter(momid == 529897525)


```


## Cox model

```{r}

library(survival)
library(survminer)

head(dat_con)

recode_fx <- function(x) {
  x = case_when(x == 10 ~ "1",
                x == 0 ~ "rm")
  x = ifelse(is.na(x), "0", x)
  return(x)
}

dat_10 <- dat_con |>
  mutate(across(week_1:week_45, ~ recode_fx(.)))

# for(j in 1:33) {
#     j = j + 5
#     
#     dat_10 <- dat_10 |>
#       filter(dat_10[, j] != "rm")
# }

dat_10 <- dat_10 |>
  mutate(across(week_1:week_44, ~relevel(as.factor(.), ref = "0")))
# week 45 has no = 0

table(dat_10$week_32)

```



```{r}

# glm(gestation ~ week_11*week_12, data = dat_10)
# 
# # null model
# res.cox <- coxph(Surv(gestation, preterm_code) ~ 
#                    1, data = dat_10)
# 
# summary(res.cox)
# 
# km <- list()
# 
# km[[1]] <- ggsurvplot(survfit(res.cox, data = dat_10), palette = "#2E9FDF",
#                       ggtheme = theme_minimal(), ylim = c(0.75, 1))
# 
# # with confounders
# res.cox_c <- coxph(Surv(gestation, preterm_code) ~ poverty_rate_sc + 
#                      momage + momeducation_char + race_ethn, 
#                    data = dat_10)
# 
# summary(res.surv_c)
# 
# km[[2]] <- ggsurvplot(survfit(res.cox_c, data = dat_10), palette = "purple",
#                       ggtheme = theme_minimal(), ylim = c(0.75, 1))
# 
# # plot the null and confounder model together
# arrange_ggsurvplots(km)
# ?arrange_ggsurvplots
# 
# # survival curve with just one confounder, to plot
# res.surv_c <- survfit(Surv(gestation, preterm_code) ~ race_ethn, 
#                       data = dat_10)
# 
# # without CI
# ggsurvplot(res.surv_c, data = dat_10, palette = "Dark2",
#            ggtheme = theme_minimal(), 
#            ylim = c(0.93, 1))
# # with CI, just two groups
# dat_f <- dat_10 |>
#   filter(race_ethn %in% c("white", "nhopi"))
# survfit(Surv(gestation, preterm_code) ~ race_ethn, data = dat_f) |>
#   ggsurvplot(data = dat_10, palette = "Dark2",
#            ggtheme = theme_minimal(), 
#            ylim = c(0.93, 1), conf.int = TRUE)
# 
# 
# ?ggsurvplot
# 
# ?coxph
# 
# 
# ?coxph
# 
# 
# summary(dat_10)
# 
# res.cox(gestation)

```

## aside: gestational hypertension

```{r}


# 
# # with confounders
# res.cox_c <- coxph(Surv(gestation, preterm_code) ~ 
#                      week_1 + week_2 + week_3 +
#                      week_4 + week_5 + week_6 +
#                      week_7 + week_8 + week_9 +
#                      week_10 + week_11 + week_12 +
#                      week_13 + week_14 + week_15 +
#                      week_16 + week_17 + week_18 +
#                      week_19 + week_20 + week_21 +
#                      week_22 + week_23 + week_24 +
#                      week_25 + week_26 + week_27 +
#                      week_28 + week_29 + week_30 +
#                      week_31 + week_32 + week_33, 
#                    data = dat_10)
# 
# summary(res.cox_c)
# 
# VIF(res.cox_c) # bit of multicollinearity
# 
# ggsurvplot(survfit(res.cox_c, data = dat_10), palette = "#2E9FDF",
#                       ggtheme = theme_minimal(), ylim = c(0.99, 1),
#            conf.int = TRUE)
# 
# survfit(Surv(gestation, preterm_code) ~ week_1 + week_2 + week_3 +
#                      week_4 + week_5 + week_6 +
#                      week_6 + week_7 + week_9 +
#                      week_10 + week_11 + week_12 +
#                      week_13 + week_14 + week_15 +
#                      week_16 + week_17 + week_18 +
#                      week_19 + week_20 + week_21 +
#                      week_22 + week_23 + week_24 +
#                      week_25 + week_26 + week_27 +
#                      week_28 + week_29 + week_30 +
#                      week_31 + week_32 + week_33, 
#                    data = dat_10) |>
#   ggsurvplot(data = dat_10, palette = "Dark2",
#            ggtheme = theme_minimal(), 
#            ylim = c(0.90, 1), conf.int = TRUE)
# 
# summary(dat_con)


```

## trying a glm with time-to-event structure

```{r}

library(tidyr)
library(stringr)
library(performance)


dat_10_long <- dat_10 |>
  pivot_longer(week_1:week_45, names_to = "exposure_week", 
               values_to = "cluster_10") |>
  # remove "rm", as the exposure does not exist (exposure occurs after event)
  filter(cluster_10 != "rm")

head(dat_10_long)


test_2 <- dat_10_long |>
  mutate(week = as.numeric(
    str_replace(exposure_week, "week_", ""))) |>
  filter(!week %in% as.character(c(34:45)))

# only exposures prior to the outcome
test_2 <- test_2 |>
  filter(week <= gestation)


# test_2[which(test_2$week <= test_2$gestation), ]

# library(lme4)
library(INLA)

test_2$momid <- as.factor(test_2$momid)

test_2$race_ethn <- relevel(as.factor(test_2$race_ethn), ref = "white")

system.time(
  res_0 <- inla(preterm_code ~ #f(momid, model = "iid", hyper = prec.prior) +
                  cluster_10*exposure_week +
                  f(week, model = "rw1") +
                  poverty_rate_sc +
                  # momage + momeducation_char + 
                  race_ethn,
                data = test_2, family = "binomial",
                control.predictor = list(compute = T),
                control.compute = list(dic = TRUE, waic = TRUE),
                control.inla = list(int.strategy = "eb"),
                # control.fixed = list(expand.factor.strategy = "inla"),
                verbose = FALSE)
)

summary(res_0)

# cluster_10*exposure_week + f(week)
# WAIC: 270802.42
# cluster_10*exposure_week + f(week) + confounders
# WAIC: 267276.57
# cluster_10*exposure_week + f(week) + only poverty and race, ethnicity (minimal adjustment set)
# WAIC: 270354.38

summary(res_1)

# plotting hyperparameters
plot(res_1, plot.fixed.effects = FALSE,
     plot.random.effects = FALSE,
     plot.hyperparameters = TRUE,
     plot.predictor = FALSE, cex = 1.25)
# not ideal, but

# taking a peak
inla.pmarginal(0, res_1$marginals.fixed$`cluster_101:exposure_weekweek_10`)
inla.pmarginal(0, res_1$marginals.fixed$`cluster_101:exposure_weekweek_11`)
inla.pmarginal(0, res_1$marginals.fixed$`cluster_101:exposure_weekweek_12`)

# storing results
results <- data.frame(variable = rownames(res_1$summary.fixed),
                      mean = res_1$summary.fixed$mean,
                      q0.025 = res_1$summary.fixed$`0.025quant`,
                      q0.975 = res_1$summary.fixed$`0.975quant`)

p_value_function <- function(x) inla.pmarginal(0, x)

results$p_value <- map_df(lapply(res_1$marginals.fixed, 
                                 p_value_function),
                          ~as.data.frame(.)) |>
  rename(p_value = `.`)

```


### plot results

```{r}

# only effects of interest; those related to exposure to cluster 10
results <- results |>
  filter(grepl("cluster_101", variable) == TRUE)

# tidying
results$variable <- str_replace(results$variable, "cluster_101:exposure_week", "")
results$variable <- ifelse(results$variable == "cluster_101", "week_1",
                           results$variable)
results$week <- as.numeric(
  str_replace(results$variable, "week_", ""))

# bonferroni adjustment
alpha_level = 0.05 / (33 * 12) # 33 weeks, 12 clusters

results$bonferroni <- ifelse(results$p_value > alpha_level, "not", "significant")
table(results$bonferroni)
# not significant, but that's alright; I've always wanted to publish a paper without any "significant" results

# exponentiating
results$or <- exp(results$mean)
results$lwr <- exp(results$q0.025)
results$upr <- exp(results$q0.975)

ggplot(results, aes(x = week, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  # facet_wrap(~ cluster, scales = "fixed") +
  ggtitle("Cluster 10 Exposure") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Exposure Week") +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1))





```



```{r}

### bonferroni
inla.hpdmarginal(0.95, res_1$marginals.fixed$`cluster_101:exposure_weekweek_10`)
 # * 396
inla.pmarginal(0.95, res_1$marginals.fixed$`cluster_101:exposure_weekweek_10`), normalize = TRUE, len = 1024)


str(data.frame(rownames(data.frame(test_glm_c$coefficients))))

VIF(glm(preterm_code ~ cluster_10 + exposure_week,
                data = test_2,
                family = "binomial"))
# no multicollinearity; GVIF < 1.004 


VIF(test_glm)

table(test$exposure_week)

summary(test_glm)
VIF(test_glm)

```






## simple cluster statistics

```{r}

```
















## cox it with time-varying exposure
- leave it be  6/5/25

```{r}

# install.packages("timereg")
# library(timereg)

# fit_time <- timecox(Surv(gestation, preterm) ~ const(poverty_rate_sc) + 
#                   const(momage) + const(momeducation_char) + const(race_ethn),
#                   max.time = 33, data = dat_10)

# start-stop format for time-varying covariates
# according to: https://stats.oarc.ucla.edu/wp-content/uploads/2025/02/survival_r_full.html
# - start and stop are time variables
#     - note that the left time point is exclusive and the right time point is inclusive
#     - i.e., A = (0, 3], B = (3, 4]
#     - This means the status is A for time steps 1, 2, 3 and B for step 4
# - event is the status variable
# each individual will have as many rows as they have changes in status + 1 (i.e, cluster assignment)


# sound easy? well try restructuring the data
tv_test <- test_2[, c("momid", "grid_id", "exposure_week", "cluster_10")]


tv_list <- list()

start_time <- Sys.time()

for(i in 1:length(unique(tv_test$momid))) {
  
  ids <- unique(tv_test$momid)
  
  sm_tv <- tv_test |>
    filter(momid == ids[i])
  
  sm_rle <- rle(sm_tv$cluster_10)
  
  sm_rle <- cbind(sm_tv[length(sm_rle[[1]]), # number of status changes + 1
                        c("momid", "grid_id")], # maternal characteristics
                  int = sm_rle[[1]], # interval length
                  status = sm_rle[[2]]) # status
  
  sm_rle$start <- shift(cumsum(sm_rle$int), type = "lag")
  sm_rle$end <- cumsum(sm_rle$int)
  sm_rle[1, "start"] <- 0
  
  sm_rle <- sm_rle |>
    select(!int)
  
  tv_list[[i]] <- sm_rle
  
}

end_time <- Sys.time()
end_time - start_time
# not so bad; a bit time intensive with the loop (53m) but works

tv_dat <- map_df(tv_list, ~as.data.frame(.))

write.csv(tv_dat, "data/exposure_start_stop.csv", row.names = FALSE)

# format is Surv(time, time2, event)

tv_dat[which(tv_dat$start > 100), ]
# well.. something went wrong
# there were subsequent pregnancies — removed in a chunk above

tv_dat$status <- as.numeric(tv_dat$status)

## add outcome
tv_comb <- merge(tv_dat, dat_10, by = c("momid", "grid_id"))

summary(tv_comb)


tv_cox_0 <- survfit(Surv(start, end, preterm_code) ~ status, 
                    data = tv_comb)

summary(tv_cox_0)
print(tv_cox_0)

ggsurvplot(tv_cox_0, conf.int = T, ylim = c(0.9, 1), xlab = "Exposure Timing")
### quite certain that this violates the proportional hazards assumption — the effect of exposure won't be constant throughout pregnancy

# plot(tv_cox_0, ylab="survival probability", xlab="weeks",
#      xlim = c(0, 50), col = c("red", "green"), yla)

```

### make this more computationally efficient
- function to convert each observation to start-stop format  
- lapply to wide dataset  

```{r}

# tv_test <- test_2[, c("momid", "grid_id", "exposure_week", "cluster_10")]

tv_sample <- dat_10[1, ]

# tv_list_sample <- list()

tv_fun <- function(dat) {
   
  # ids <- unique(dat$momid)
  
  # sm_tv <- dat |>
  #   filter(momid == ids)
  
  long <- dat |>
    pivot_longer(week_1:week_45, names_to = "exposure_week", 
               values_to = "cluster_10") |>
  # remove "rm", as the exposure does not exist (exposure occurs after event)
    filter(cluster_10 != "rm") |>
    mutate(week = as.numeric(
      str_replace(exposure_week, "week_", ""))) |>
    filter(!week %in% as.character(c(34:45))) |>
    filter(week <= gestation) |> # only exposures prior to the outcome
    mutate(momid = as.factor(momid))
 
  
  sm_rle <- rle(long$cluster_10)
  return(sm_rle)
  sm_rle <- cbind(sm_tv[length(sm_rle[[1]]), # number of status changes + 1
                        c("momid", "grid_id")], # maternal characteristics
                  int = sm_rle[[1]], # interval length
                  status = sm_rle[[2]]) # status
  
  sm_rle$start <- shift(cumsum(sm_rle$int), type = "lag")
  sm_rle$end <- cumsum(sm_rle$int)
  sm_rle[1, "start"] <- 0
  
  sm_rle <- sm_rle |>
    select(!int)
  
  return(sm_rle)
  
}

tv_fun(tv_sample)


tv_list_2 <- list()

start_time <- Sys.time()

for(i in 1:length(unique(tv_test$momid))) {
  
  ids <- unique(tv_test$momid)
  
  sm_tv <- tv_test |>
    filter(momid == ids[i])
  
  sm_rle <- rle(sm_tv$cluster_10)
  
  sm_rle <- cbind(sm_tv[length(sm_rle[[1]]), # number of status changes + 1
                        c("momid", "grid_id")], # maternal characteristics
                  int = sm_rle[[1]], # interval length
                  status = sm_rle[[2]]) # status
  
  sm_rle$start <- shift(cumsum(sm_rle$int), type = "lag")
  sm_rle$end <- cumsum(sm_rle$int)
  sm_rle[1, "start"] <- 0
  
  sm_rle <- sm_rle |>
    select(!int)
  
  tv_list[[i]] <- sm_rle
  
}

end_time <- Sys.time()
end_time - start_time
# not so bad; a bit time intensive with the loop (53m) but works

tv_dat <- map_df(tv_list, ~as.data.frame(.))

```
