==> 03_model loop.Rmd <==
---
title: "03_model loop"
author: "Brenna Kelly"
date: "2025-06-06"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r}

library(INLA)
library(tidyr)
library(dplyr)
library(purrr)
library(stringr)
library(survival)
library(regclass)
library(survminer)
library(data.table)
library(performance)

# OG
full_dat <- read.csv("data/clean_birth_data.csv")
# outcome: placental insufficiency
full_dat$hypertension_code <- ifelse(full_dat$hypertension %in% c("C", "E", "H", "P"),
                                     "hyp", "no")
full_dat$pl_insuf <- case_when(full_dat$sga10_by_sex > full_dat$birthweightgrams ~ 1, # sga
                               full_dat$hypertension %in% c("H", "E", "P") ~ 1)
full_dat$pl_insuf <- ifelse(is.na(full_dat$pl_insuf), 0, 1)

# outcome: gestational hypertension
full_dat$gestational_hyp <- ifelse(full_dat$hypertension %in% c("H", "E", "P"), 1, 0)

# outcome: gestational diabetes
full_dat$gestational_diab <- ifelse(full_dat$gestationaldiabetes == 1, 1, 0)

# outcome: preterm birth
full_dat$preterm <- ifelse(full_dat$gestation < 34, "preterm", "term")
full_dat$preterm_code <- ifelse(full_dat$preterm == "preterm", 1, 0)

# outcome: very preterm birth
full_dat$very_preterm <- ifelse(full_dat$gestation < 30, "preterm", "term")

# *negative control* outcome: gender
full_dat$gender_code <- ifelse(full_dat$gender == "M", 1, 0)

# grab confounders
dat <- fread("/Users/brenna/Documents/School/Research/distributed_lag/data/birth_exposures_with_confounders.csv")

confounders <- dat[, c("momid", "grid_id", "gestation", "birthccyy", "birthmm", 
                       "mommedicaid_char", "mommaritalstatus_char",
                       "momwic_char", "momeducation_char", "momage",
                       "poverty_rate")]
rm(dat)

# and a few more
# mom <- read.csv('../../data/health data/births.csv')
mom_2013 <- fread('/Users/brenna/Downloads/births 2013+.csv')
mom_2016 <- fread('/Users/brenna/Downloads/births 2016+.csv')

mom <- rbind(mom_2013, mom_2016)

mom <- mom |>
  mutate(MomHeightFeet = as.numeric(MomHeightFeet),
         MomHeightFeet = ifelse(MomHeightFeet == 1, NA, MomHeightFeet),
         MomHeightInches = as.numeric(MomHeightInches),
         race_ethn = case_when(MomHispanicOrigin == 1 | MomHispanicMex == 1 |
                                 MomHispanicCuban == 1 | MomHispanicPuertoRican == 1 |
                                 MomHispanicOther == 1 | MomHispanicOtherSpec != "" |
                                 !MomHispanicOtherSpec %in% c("*", "") ~ # primacy
                                 "hispanic",
                               MomRaceWhite == 1 ~ "white",
                               MomRaceBlack == 1 ~ "black",
                               MomRaceAmIndian == 1 ~ "amindian",
                               MomRaceChinese == 1 | MomRaceJapanese == 1 |
                                 MomRaceFilipino == 1 | MomRaceOtherAsian == 1 |
                                 MomRaceAsianIndian == 1 | MomRaceKorean == 1 |
                                 MomRaceVietnamese == 1 ~ "asian",
                               MomRaceHawaiian == 1 | MomRaceSamoan == 1 |
                                 MomRaceGuamanian == 1 | MomRacePacIslander == 1 |
                                 MomRaceTongan == 1 ~ "nhopi",
                               MomRaceOther == 1 | MomRaceOtherSpec1 == 1 |
                                 MomRaceOtherSpec2 == 1 | MomRaceUnknown == 1 ~ "other"),
         race_ethn = ifelse(is.na(race_ethn), "other", race_ethn),
         # there's one maternal_wt = 550; 
         # this could be a recording error, but it's prior to 2013 so don't bother with it
         MomWeightPrior = as.numeric(MomWeightPrior),
         # bmi = ((lbs * 703) / inches) / inches
         mom_bmi = ((MomWeightPrior * 703) / 
                      (12 * MomHeightFeet + MomHeightInches)) / 
           (12 * MomHeightFeet + MomHeightInches))

names(mom) <- tolower(names(mom))
mom <- mom[, c("momid", "birthccyy", "birthmm",
               "race_ethn")] #, "mom_bmi", "momheightfeet",
#"momheightinches", "momweightprior"
# note: too much missing in BMI
mom$momid <- as.numeric(mom$momid)

# and a couple more
table(full_dat[, c("prevpregnowliving", "prevpregnowdead", "prevpregstillborn")])
# none — these were excluded; our cohort is nulligravida (as far as we know)

# birth data
full_dat <- merge(full_dat, mom, by = c("momid", "birthccyy", "birthmm"))
z <- read.csv("data/birth_traj_trunc_jun08.csv")

z_y <- merge(z, full_dat[, c("momid", "grid_id", "birthccyy", "birthmm",
                             "sga", "hypertension_code",
                             "pl_insuf", "preterm", "preterm_code",
                             "gestational_hyp", "very_preterm",
                             "gestational_diab", "gender_code",
                             "race_ethn")], 
             by = c("momid", "grid_id")) # N = 44984

dat_con <- merge(z_y, confounders, by = c("momid", "grid_id", 
                                          "birthccyy", "birthmm"))

dat_con$poverty_rate <- ifelse(is.na(dat_con$poverty_rate), 0, dat_con$poverty_rate)
dat_con$poverty_rate_sc <- (dat_con$poverty_rate - mean(dat_con$poverty_rate)) / 10

# drop subsequent pregnancies
test <- dat_con |>
  arrange(momid, birthccyy, birthmm)

dat_con <- test[match(unique(test$momid), test$momid), ]



```

### Table 1

```{r}

library(tableone)

# find non-normal variables
hist(dat_con$momage)

?CreateTableOne

tab1 <- print(
  CreateTableOne(c("race_ethn", "momage", "poverty_rate", "gestation", "birthccyy"),
                 strata = "preterm", data = dat_con),
  quote = FALSE, noSpaces = TRUE, printToggle = FALSE,
  nonnormal = c("momage", "gestation"))

# tab3Mat <- print(tab3, nonnormal = biomarkers, exact = "stage", )
## Save to a CSV file
write.csv(tab1, file = "data/results/table_1_mom.csv")


```


```{r}


inla_models <- list()
inla_results <- list()

for(i in c(3, 4, 7, 8)) {
  
  ##### encoding clusters
  recode_fx <- function(x) {
    x = case_when(x == i ~ "1",
                  x == 0 ~ "rm")
    x = ifelse(is.na(x), "0", x)
    return(x)
  }
  
  dat_10 <- dat_con |>
    mutate(across(week_1:week_45, ~ recode_fx(.)))
  
  dat_10 <- dat_10 |>
    mutate(across(week_1:week_44, ~relevel(as.factor(.), ref = "0")))
  # week 45 has no = 0
  
  ##### reformatting
  dat_10_long <- dat_10 |>
    pivot_longer(week_1:week_45, names_to = "exposure_week", 
                 values_to = "cluster_10") |>
    # remove "rm", as the exposure does not exist (exposure occurs after event)
    filter(cluster_10 != "rm") |>
    mutate(week = as.numeric(
      str_replace(exposure_week, "week_", ""))) |>
    filter(!week %in% as.character(c(34:45))) |>
    filter(week <= gestation)

  
  dat_10_long$momid <- as.factor(dat_10_long$momid)
  
  dat_10_long$race_ethn <- relevel(as.factor(dat_10_long$race_ethn), ref = "white")
  
  summary(dat_10_long$cluster_10)
  
  res_1 <- inla(preterm_code ~ #f(momid, model = "iid", hyper = prec.prior) +
                  cluster_10*exposure_week +
                  f(week, model = "rw1") +
                  poverty_rate_sc +
                  # momage + momeducation_char +
                  race_ethn,
                data = dat_10_long, family = "binomial",
                control.predictor = list(compute = T),
                control.compute = list(dic = TRUE, waic = TRUE),
                control.inla = list(int.strategy = "eb"),
                # control.fixed = list(expand.factor.strategy = "inla"),
                verbose = FALSE)
  
  # print(summary(res_1))
  
  inla_models[[i]] <- summary(res_1)

  # storing results
  results <- data.frame(variable = rownames(res_1$summary.fixed),
                        mean = res_1$summary.fixed$mean,
                        q0.025 = res_1$summary.fixed$`0.025quant`,
                        q0.975 = res_1$summary.fixed$`0.975quant`)
  
  p_value_function <- function(x) inla.pmarginal(0, x)
  
  results$p_value <- map_df(lapply(res_1$marginals.fixed, 
                                   p_value_function),
                            ~as.data.frame(.)) |>
    rename(p_value = `.`)
  
  results <- results |>
    mutate(p_value = case_when(results$mean < 0 ~ (1 - results$p_value),
                               results$mean > 0 ~ results$p_value))
  
  # results$p_value <- ifelse(results$mean < 0,
  #                           (1 - results$p_value),
  #                           results$p_value)
  # .. the strangest thing; ^this doesn't work
  
  results <- results |>
    filter(grepl("cluster_101", variable) == TRUE)
  
  # tidying
  results$variable <- str_replace(results$variable, "cluster_101:exposure_week", "")
  results$variable <- ifelse(results$variable == "cluster_101", "week_1",
                             results$variable)
  results$week <- as.numeric(
    str_replace(results$variable, "week_", ""))
  
  # bonferroni adjustment
  alpha_level = 0.05 / (33 * 12) # 33 weeks, 12 clusters
  
  # results$bonferroni <- ifelse(results$p_value > alpha_level, "not", "significant")
  # table(results$bonferroni)
  # not significant, but that's alright; I've always wanted to publish a paper without any "significant" results
  
  # exponentiating
  results$or <- exp(results$mean)
  results$lwr <- exp(results$q0.025)
  results$upr <- exp(results$q0.975)
  
  results$cluster <- i
  
  inla_results[[i]] <- results
  
  print(paste("finished with cluster", i))
  
  # write.csv(paste0("data/results/cluster",
  #                  str_pad(i, width = 2, side = "left", pad = "0"),
  #                  ".csv"), row.names = FALSE)

}

# cl_1 <- inla_models
# inla_models

inla_models
inla_results

```


```{r}
### plotting hyperparameters
# 
# 
# for(i in inla_models) {
#   
#   return(plot(res_1, plot.fixed.effects = FALSE,
#               plot.random.effects = FALSE,
#               plot.hyperparameters = TRUE,
#               plot.predictor = FALSE, cex = 1.25))
#   
# }
#  
```



```{r}

## plot of all ORs

cl_results <- map_df(inla_results, ~as.data.frame(.))

cl_results$cluster <- as.factor(cl_results$cluster)

# multiple comparisons with false discovery rate
cl_results$p_fdr <- p.adjust(unlist(cl_results$p_value), method = "fdr")
cl_results$p_value <- cl_results$p_value$p_value
cl_results <- cl_results |>
  arrange(cluster, week)

# write.csv(cl_results, "data/results/inla_results.csv", row.names = FALSE)
# don't duplicate information (written below)

cl_results$sig <- ifelse(cl_results$p_fdr < 0.05, "significant", "not")

write.csv(cl_results, "data/results/cl_results.csv", row.names = FALSE)

# plot all results, by FDR significance
cols <- c("significant" = "red", "not" = "black")

library(extrafont)

cl_plt <- ggplot(cl_results, aes(x = week, y = or, colour = sig)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  scale_colour_manual(values = cols, labels = c("p < 0.05 after FDR", "p ≥ 0.05 after FDR")) +
  # facet_wrap(~ cluster, scales = "fixed") +
  ggtitle("Exposure to Mixtures (Clusters) Throughout Pregnancy") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Exposure Week") +
  # ylim(0.15, 5) +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.5, hjust = 1)) +
  facet_wrap(~cluster, scales = "free") +
  theme(text = element_text(colour = "black",
                            family = "Century Gothic",
                            face = "plain"))

## export

cl_plt

class(cl_results)
##### just cluster 5, 10, 1
cl_results |>
  filter(cluster %in% c(1, 5, 10, 11)) |>
  ggplot(aes(x = week, y = or, colour = sig)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  scale_colour_manual(values = c("black", "red")) +
  # facet_wrap(~ cluster, scales = "fixed") +
  ggtitle("Exposure to Mixtures (Clusters) Throughout Pregnancy") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Exposure Week") +
  # ylim(0.15, 5) +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.5, hjust = 1)) +
  facet_wrap(~cluster, scales = "fixed")



## what is the critical window for cluster 5?
cl_results |>
  filter(cluster == 5) |>
  arrange(week)
# ~ weeks 24-29
## what is the critical window for cluster 10?
cl_results |>
  filter(cluster == 10) |>
  arrange(week)
# ~ weeks 9-13
### use this for the duration models

## does cluster 1 have a critical window?
cl_results |>
  filter(cluster == 1) |>
  arrange(week) |>
  mutate(p_value = 1 - p_value)
# - weeks 9-13
## how about cluster 11?
cl_results |>
  filter(cluster == 11) |>
  arrange(week) |>
  mutate(p_value = 1 - p_value)
## this does not even approach significance

```


```{r}

## are there significant (ala bonferroni) results anywhere?
table(cl_results$bonferroni)
# there's just one

# cluster 5, gestation week 28
cl_results[which(cl_results$bonferroni == "significant"), ]

```

### levels of preterm  
Late preterm (>37 weeks) is more likely to be medically indicated. Let's see if there is an association between cluster 5, which is associated with increased risk of preterm birth when exposed in week 28, and varying levels of preterm. Also comparing with cluster 10, which is associated with increased risk of preterm birth during late-first and early-second trimesters.  
***don't go down this rabbithole***

```{r}

dat_con$ptb_class <- case_when(dat_con$gestation <= 28 ~ "early preterm",
                               dat_con$gestation <= 34 ~ "preterm",
                               dat_con$gestation <= 37 ~ "late preterm",
                               dat_con$gestation > 37 ~ "term")

table(dat_con$ptb_class, dat_con$week_28)
# first thought — it looks like late preterm is more common everywhere, but maybe (?) more common with cluster 5 exposure in week 28
#     just a note, in case we forgot — pregnancies are only include up until birth
prop.table(table(dat_con$ptb_class, dat_con$week_28), marg = 2)
# looks like late preterm birth is (roughly) just as likely when exposed to cluster 5 as any other cluster during week 28 
# the thought was — is exposure to cluster 5 later in pregnancy caused by different etiology than exposure to cluster 10 earlier in pregnancy

# maybe — is it the combination? is exposure to cluster 10 early associated with exposure to cluster 5 later
# looking at the plot of exposure over time — it's highly possible, as cluster 5 tends to follow cluster 10 throughout the year

```




```{r}

dat_con$week_11_exp <- relevel(as.factor(
  ifelse(dat_con$week_11 == 10, "ten", "not ten")), ref = "not ten")
dat_con$week_28_exp <- relevel(as.factor(
  ifelse(dat_con$week_28 == 5, "five", "not five")), ref = "not five")

test <- inla(preterm_code ~ week_11_exp*week_28_exp +
               poverty_rate_sc + race_ethn, 
             data = dat_long_10, family = "binomial")

# table(dat_con$week)

p_value_function(test$marginals.fixed$`week_11_expten:week_28_expfive`)

summary(test)
# it looks — maybe — like the combination of 5 and 10 has a large effect, but when we model their combination, 5 does not matter; the effect of 10 is high but not the same as the combination of 10 and 5

# ~ effect of week 11 cluster 10 (maybe.. if I did this right)
exp(0.219)
p_value_function(test$marginals.fixed$`week_11_expten`) # definitely not significant at 0.0001 or whatever

```


# maybe — make the referent cluster 6; and try everything again, just beginning with cluster 10

```{r}


library(survival)
library(survminer)

head(dat_con)

recode_fx <- function(x) {
  x = case_when(#x == 10 ~ "1",
                x == 0 ~ "rm")
  # x = ifelse(is.na(x), "0", x)
  return(x)
}

# dat_10 <- dat_con |>
#   mutate(across(week_1:week_45, ~ recode_fx(.)))

head(dat_10)
# for(j in 1:33) {
#     j = j + 5
# 
#     dat_10 <- dat_10 |>
#       filter(dat_10[, j] != "rm")
# }

dat_6 <- dat_con |>
  mutate(across(week_1:week_43, ~relevel(as.factor(.), ref = "6")))
# week 45 has no = 0

summary(dat_6)



library(tidyr)
library(stringr)
library(performance)


dat_6_long <- dat_6 |>
  pivot_longer(week_1:week_43, names_to = "exposure_week", 
               values_to = "cluster") |>
  # remove "rm", as the exposure does not exist (exposure occurs after event)
  filter(cluster != "rm")

head(dat_6_long)


test_6 <- dat_6_long |>
  mutate(week = as.numeric(
    str_replace(exposure_week, "week_", ""))) |>
  filter(!week %in% as.character(c(34:45)))

# only exposures prior to the outcome
test_6 <- test_6 |>
  filter(week <= gestation)


# test_2[which(test_2$week <= test_2$gestation), ]

# library(lme4)
# library(INLA)

test_6$momid <- as.factor(test_6$momid)

test_6$race_ethn <- relevel(as.factor(test_6$race_ethn), ref = "white")

system.time(
  res_6 <- inla(preterm_code ~ #f(momid, model = "iid", hyper = prec.prior) +
                  cluster*exposure_week +
                  f(week, model = "rw1") +
                  poverty_rate_sc +
                  # momage + momeducation_char + 
                  race_ethn,
                data = test_6, family = "binomial",
                control.predictor = list(compute = T),
                control.compute = list(dic = TRUE, waic = TRUE),
                control.inla = list(int.strategy = "eb"),
                # control.fixed = list(expand.factor.strategy = "inla"),
                verbose = FALSE)
)

summary(res_6)





#####
# storing results
results <- data.frame(variable = rownames(res_6$summary.fixed),
                      mean = res_6$summary.fixed$mean,
                      q0.025 = res_6$summary.fixed$`0.025quant`,
                      q0.975 = res_6$summary.fixed$`0.975quant`)

p_value_function <- function(x) inla.pmarginal(0, x)

results$p_value <- map_df(lapply(res_6$marginals.fixed, 
                                 p_value_function),
                          ~as.data.frame(.)) |>
  rename(p_value = `.`)

head(results)


results_10 <- results# |>
    # filter(grepl("cluster10", variable) == TRUE)

# tidying
# results_10$variable <- str_replace(results_10$variable, "cluster10:exposure_week", "")
results_10$exposure <- str_split_fixed(results_10$variable, ":", n = 2)[, 1]
results_10$week <- str_split_fixed(results_10$variable, ":", n = 2)[, 2]
# results_10$variable <- ifelse(results_10$variable == "cluster10", "week_1",
#                            results_10$variable)
results_10$week <- as.numeric(
  str_replace(results_10$week, "exposure_weekweek_", ""))

head(results_10)

# bonferroni adjustment
alpha_level = 0.05 / (33 * 12) # 33 weeks, 12 clusters

results_10$bonferroni <- ifelse(results_10$p_value > alpha_level, "not", "significant")
table(results_10$bonferroni)
# not significant, but that's alright; I've always wanted to publish a paper without any "significant" results

# exponentiating
results_10$or <- exp(results_10$mean)
results_10$lwr <- exp(results_10$q0.025)
results_10$upr <- exp(results_10$q0.975)

results_10 <- results_10 |>
  filter(grepl("cluster", exposure) == TRUE) |>
  mutate(exposure = str_replace(exposure, "cluster", ""))

results_10 <- results_10 |>
  filter(exposure != 0) |>
  mutate(exposure = as.numeric(exposure))

ref_allelse <- ggplot(results_10, aes(x = week, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  # facet_wrap(~ cluster, scales = "fixed") +
  ggtitle("Cluster 10, ref: all") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Exposure Week") +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1)) +
  facet_wrap(~exposure, scales = "free")
ref_allelse

ref_6 <- cl_results |>
  filter(cluster == 5) |>
  ggplot(aes(x = week, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  # facet_wrap(~ cluster, scales = "fixed") +
  ggtitle("Cluster 10, ref: 6") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Exposure Week") +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1))


ggarrange(ref_allelse, ref_6)

### storing results
write.csv(results_10, "inla_compared to 6.csv", row.names = FALSE)



```
