==> 04_duration model.Rmd <==
---
title: "04_duration model"
author: "Brenna Kelly"
date: "2025-06-07"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}

library(INLA)
library(tidyr)
library(dplyr)
library(purrr)
library(stringr)
library(survival)
library(regclass)
library(survminer)
library(data.table)
library(performance)

# OG
full_dat <- read.csv("data/clean_birth_data.csv")
# outcome: placental insufficiency
full_dat$hypertension_code <- ifelse(full_dat$hypertension %in% c("C", "E", "H", "P"),
                                     "hyp", "no")
full_dat$pl_insuf <- case_when(full_dat$sga10_by_sex > full_dat$birthweightgrams ~ 1, # sga
                               full_dat$hypertension %in% c("H", "E", "P") ~ 1)
full_dat$pl_insuf <- ifelse(is.na(full_dat$pl_insuf), 0, 1)

# outcome: gestational hypertension
full_dat$gestational_hyp <- ifelse(full_dat$hypertension %in% c("H", "E", "P"), 1, 0)

# outcome: gestational diabetes
full_dat$gestational_diab <- ifelse(full_dat$gestationaldiabetes == 1, 1, 0)

# outcome: preterm birth
full_dat$preterm <- ifelse(full_dat$gestation < 34, "preterm", "term")
full_dat$preterm_code <- ifelse(full_dat$preterm == "preterm", 1, 0)

# outcome: very preterm birth
full_dat$very_preterm <- ifelse(full_dat$gestation < 30, "preterm", "term")

# *negative control* outcome: gender
full_dat$gender_code <- ifelse(full_dat$gender == "M", 1, 0)

# grab confounders
dat <- fread("/Users/brenna/Documents/School/Research/distributed_lag/data/birth_exposures_with_confounders.csv")

confounders <- dat[, c("momid", "grid_id", "gestation", "birthccyy", "birthmm", 
                       "mommedicaid_char", "mommaritalstatus_char",
                       "momwic_char", "momeducation_char", "momage",
                       "poverty_rate")]
rm(dat)

# and a few more
mom <- read.csv('../../data/health data/births.csv')
mom_2013 <- fread('/Users/brenna/Downloads/births 2013+.csv')
mom_2016 <- fread('/Users/brenna/Downloads/births 2016+.csv')

mom <- rbind(mom_2013, mom_2016)

mom <- mom |>
  mutate(MomHeightFeet = as.numeric(MomHeightFeet),
         MomHeightFeet = ifelse(MomHeightFeet == 1, NA, MomHeightFeet),
         MomHeightInches = as.numeric(MomHeightInches),
         race_ethn = case_when(MomHispanicOrigin == 1 | MomHispanicMex == 1 |
                                 MomHispanicCuban == 1 | MomHispanicPuertoRican == 1 |
                                 MomHispanicOther == 1 | MomHispanicOtherSpec != "" |
                                 !MomHispanicOtherSpec %in% c("*", "") ~ # primacy
                                 "hispanic",
                               MomRaceWhite == 1 ~ "white",
                               MomRaceBlack == 1 ~ "black",
                               MomRaceAmIndian == 1 ~ "amindian",
                               MomRaceChinese == 1 | MomRaceJapanese == 1 |
                                 MomRaceFilipino == 1 | MomRaceOtherAsian == 1 |
                                 MomRaceAsianIndian == 1 | MomRaceKorean == 1 |
                                 MomRaceVietnamese == 1 ~ "asian",
                               MomRaceHawaiian == 1 | MomRaceSamoan == 1 |
                                 MomRaceGuamanian == 1 | MomRacePacIslander == 1 |
                                 MomRaceTongan == 1 ~ "nhopi",
                               MomRaceOther == 1 | MomRaceOtherSpec1 == 1 |
                                 MomRaceOtherSpec2 == 1 | MomRaceUnknown == 1 ~ "other"),
         race_ethn = ifelse(is.na(race_ethn), "other", race_ethn),
         # there's one maternal_wt = 550; 
         # this could be a recording error, but it's prior to 2013 so don't bother with it
         MomWeightPrior = as.numeric(MomWeightPrior),
         # bmi = ((lbs * 703) / inches) / inches
         mom_bmi = ((MomWeightPrior * 703) / 
                      (12 * MomHeightFeet + MomHeightInches)) / 
           (12 * MomHeightFeet + MomHeightInches))

names(mom) <- tolower(names(mom))
mom <- mom[, c("momid", "birthccyy", "birthmm",
               "race_ethn")] #, "mom_bmi", "momheightfeet",
#"momheightinches", "momweightprior"
# note: too much missing in BMI
mom$momid <- as.numeric(mom$momid)

# and a couple more
table(full_dat[, c("prevpregnowliving", "prevpregnowdead", "prevpregstillborn")])
# none — these were excluded; our cohort is nulligravida (as far as we know)

# birth data
full_dat <- merge(full_dat, mom, by = c("momid", "birthccyy", "birthmm"))
z <- read.csv("data/birth_traj_trunc_jun08.csv")

z_y <- merge(z, full_dat[, c("momid", "grid_id", "birthccyy", "birthmm",
                             "sga", "hypertension_code",
                             "pl_insuf", "preterm", "preterm_code",
                             "gestational_hyp", "very_preterm",
                             "gestational_diab", "gender_code",
                             "race_ethn")], 
             by = c("momid", "grid_id")) # N = 44984

dat_con <- merge(z_y, confounders, by = c("momid", "grid_id", 
                                          "birthccyy", "birthmm"))

dat_con$poverty_rate <- ifelse(is.na(dat_con$poverty_rate), 0, dat_con$poverty_rate)
dat_con$poverty_rate_sc <- (dat_con$poverty_rate - mean(dat_con$poverty_rate)) / 10

# drop subsequent pregnancies
test <- dat_con |>
  arrange(momid, birthccyy, birthmm)

dat_con <- test[match(unique(test$momid), test$momid), ]

```



```{r}

names(dat_con)

## 6-week critical window for cluster 5: weeks 24-29
cw_5 <- dat_con[, 29:34]
## 6-week critical window for cluster 10: weeks 9-13
cw_10 <- dat_con[, 14:19]
## 6-week critical window for cluster 1: weeks 9-13
cw_1 <- dat_con[, 14:18]

# increased duration of exposure to cluster 10
dat_con$duration_10 <- apply(cw_10, 1, function(x) length(which(x==as.character("10"))))

dat_con <- dat_con |>
  mutate(duration_10 = as.factor(duration_10))

m_10 <- inla(preterm_code ~ duration_10 +
               poverty_rate_sc + race_ethn, 
             family = "binomial", data = dat_con,
             control.predictor = list(compute = T),
             control.compute = list(dic = TRUE, waic = TRUE),
             verbose = FALSE)
summary(m_10)

# increased duration of exposure to cluster 5
dat_con$duration_5 <- apply(cw_5, 1, function(x) length(which(x==as.character("5"))))

dat_con <- dat_con |>
  mutate(duration_5 = as.factor(duration_5))

m_5 <- inla(preterm_code ~ duration_5 +
              poverty_rate_sc + race_ethn, 
            family = "binomial", data = dat_con,
            control.predictor = list(compute = T),
            control.compute = list(dic = TRUE, waic = TRUE),
            verbose = FALSE)
summary(m_5)

# increased duration of exposure to cluster 1
dat_con$duration_1 <- apply(cw_1, 1, function(x) length(which(x==as.character("1"))))

dat_con <- dat_con |>
  mutate(duration_1 = as.factor(duration_1))

m_1 <- inla(preterm_code ~ duration_1 +
              poverty_rate_sc + race_ethn, 
            family = "binomial", data = dat_con,
            control.predictor = list(compute = T),
            control.compute = list(dic = TRUE, waic = TRUE),
            verbose = FALSE)
summary(m_1)

```

### Storing results  

```{r}

p_value_function <- function(x) inla.pmarginal(0, x)

## Bonferroni is dead
# duration_alpha_level = 0.05 / (6 * 12) # 6 week-window, for 12 clusters
# duration_alpha_level_1 = 0.05 / (5 * 12) # 1 week-window, for 12 clusters; for cw_1
## Long live FDR (see below, )

# storing results
cw_res_10 <- data.frame(variable = rownames(m_10$summary.fixed),
                        mean = m_10$summary.fixed$mean,
                        q0.025 = m_10$summary.fixed$`0.025quant`,
                        q0.975 = m_10$summary.fixed$`0.975quant`) |>
  mutate(p_value <- map_df(lapply(m_10$marginals.fixed, 
                                   p_value_function),
                            ~as.data.frame(.))) |>
    rename(p_value = `.`) |>
  filter(grepl("duration", variable) == TRUE) |>
  mutate(duration = as.numeric(str_replace(variable, "duration_10", ""))) |>
  mutate(or = exp(mean)) |>
  mutate(lwr = exp(q0.025)) |>
  mutate(upr = exp(q0.975))

cw_res_5 <- data.frame(variable = rownames(m_5$summary.fixed),
                       mean = m_5$summary.fixed$mean,
                       q0.025 = m_5$summary.fixed$`0.025quant`,
                       q0.975 = m_5$summary.fixed$`0.975quant`) |>
  mutate(p_value <- map_df(lapply(m_5$marginals.fixed, 
                                  p_value_function),
                           ~as.data.frame(.))) |>
  rename(p_value = `.`) |>
  filter(grepl("duration", variable) == TRUE) |>
  mutate(duration = as.numeric(str_replace(variable, "duration_5", ""))) |>
  mutate(or = exp(mean)) |>
  mutate(lwr = exp(q0.025)) |>
  mutate(upr = exp(q0.975))

cw_res_1 <- data.frame(variable = rownames(m_1$summary.fixed),
                       mean = m_1$summary.fixed$mean,
                       q0.025 = m_1$summary.fixed$`0.025quant`,
                       q0.975 = m_1$summary.fixed$`0.975quant`) |>
  mutate(p_value <- map_df(lapply(m_1$marginals.fixed, 
                                  p_value_function),
                           ~as.data.frame(.))) |>
  rename(p_value = `.`) |>
  mutate(p_value = 1 - p_value) |>
  filter(grepl("duration", variable) == TRUE) |>
  mutate(duration = as.numeric(str_replace(variable, "duration_1", ""))) |>
  mutate(or = exp(mean)) |>
  mutate(lwr = exp(q0.025)) |>
  mutate(upr = exp(q0.975))

# adjust for all comparisons — all windows
cw_res_10 <- cw_res_10 |>
  mutate(fdr = p.adjust(c(cw_res_10$p_value, cw_res_5$p_value, cw_res_1$p_value), 
                        method = "fdr")[1:6],
         fdr_sig = ifelse(fdr < 0.05, "significant", "not"))
cw_res_5 <- cw_res_5 |>
  mutate(fdr = p.adjust(c(cw_res_10$p_value, cw_res_5$p_value, cw_res_1$p_value), 
                        method = "fdr")[7:12],
         fdr_sig = ifelse(fdr < 0.05, "significant", "not"))
cw_res_1 <- cw_res_1 |>
  mutate(fdr = p.adjust(c(cw_res_10$p_value, cw_res_5$p_value, cw_res_1$p_value), 
                        method = "fdr")[13:17],
         fdr_sig = ifelse(fdr < 0.05, "significant", "not"))

# write.csv(cw_res_10, "data/results/cw_res_10.csv", row.names = FALSE)
# cw_res_10 <- read.csv("data/results/cw_res_10.csv")
# 
# write.csv(cw_res_5, "data/results/cw_res_5.csv", row.names = FALSE)
# cw_res_5 <- read.csv("data/results/cw_res_5.csv")
# 
# write.csv(cw_res_1, "data/results/cw_res_1.csv", row.names = FALSE)
# cw_res_1 <- read.csv("data/results/cw_res_1.csv")


```

### Plotting

```{r}

# plots
plt_cw_10 <- ggplot(cw_res_10, aes(x = duration, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  ggtitle("Cluster 10 Exposure \nin Weeks 9-14") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Duration of Exposure (Weeks)") +
  ylim(c(0, 4)) +
  theme_classic() +
  scale_x_continuous(breaks = 1:6) +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1))

plt_cw_5 <- ggplot(cw_res_5, aes(x = duration, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  ggtitle("Cluster 5 Exposure \nin Weeks 24-29") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Duration of Exposure (Weeks)") +
  ylim(c(0, 4)) +
  theme_classic() +
  scale_x_continuous(breaks = 1:6) +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1))

plt_cw_1 <- ggplot(cw_res_1, aes(x = duration, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  ggtitle("Cluster 1 Exposure \nin Weeks 9-13") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Duration of Exposure (Weeks)") +
  ylim(c(0, 4)) +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1))

ggarrange(plt_cw_10,
          plt_cw_5,
          plt_cw_1,
          ncol = 3)

# save as 12"x5", durations_panel.png

```

### a few stats (for the supplement?)

```{r}

library(tableone)

CreateTableOne(vars = c("duration_10", "duration_5"), strata = "preterm", data = dat_con)
# duration varies significantly across outcome for cluster 10, but much less so for cluster 5

cor.test(as.numeric(dat_con$gestation), 
         as.numeric(dat_con$duration_10))
# slight but significant negative correlation between duration of cluster 10 exposure and gestational age

table(comp_5_10$duration_10)
# how common was exposure to the full window

```


```{r}

# the combination of clsuter 5 and 10; for reviewers only, as needed

comp_5_10 <- dat_con |>
  mutate(exposure_10_cw = ifelse(duration_10 != 0, 1, 0)) |>
  mutate(exposure_5_cw = ifelse(duration_5 != 0, 1, 0))

table(comp_5_10$exposure_10_cw,
      comp_5_10$exposure_5_cw)

comparison_5_10 <- inla(preterm_code ~ exposure_10_cw*exposure_5_cw +
                          poverty_rate_sc + race_ethn, 
                        family = "binomial", data = comp_5_10,
                        control.predictor = list(compute = T),
                        control.compute = list(dic = TRUE, waic = TRUE),
                        verbose = FALSE)

summary(comparison_5_10)

# save this information in case it comes up in review
comparison_5_10 <- data.frame(variable = rownames(comparison_5_10$summary.fixed),
                       mean = comparison_5_10$summary.fixed$mean,
                       q0.025 = comparison_5_10$summary.fixed$`0.025quant`,
                       q0.975 = comparison_5_10$summary.fixed$`0.975quant`) |>
  mutate(p_value <- map_df(lapply(comparison_5_10$marginals.fixed, 
                                  p_value_function),
                           ~as.data.frame(.))) |>
  rename(p_value = `.`) |>
  # filter(grepl("duration", variable) == TRUE) |>
  # mutate(duration = as.numeric(str_replace(variable, "duration_5", ""))) |>
  mutate(or = exp(mean)) |>
  mutate(lwr = exp(q0.025)) |>
  mutate(upr = exp(q0.975)) #|>
  # bonferroni is dead
  # mutate(bonferroni = ifelse(p_value > duration_alpha_level, "not", "significant"))
# write.csv(comparison_5_10, "data/results/comparison_5_10.csv", row.names = FALSE)
# comparison_5_10 <- read.csv("data/results/comparison_5_10.csv")


```




### now do this for all other clusters, for the supplement  

Run the model for increased duration of exposure during cluster 5's critical window for all other clusters.

```{r}

## cluster 5's critical window
cw_5_loop <- list()

for(i in 1:12) {
  
  dat_con$duration_5 <- apply(cw_5, 1, function(x) length(which(x==as.character(i))))
  
  dat_con <- dat_con |>
    mutate(duration_5 = as.factor(duration_5))
  
  m_5_loop <- inla(preterm_code ~ duration_5 +
                     poverty_rate_sc + race_ethn, 
                   family = "binomial", data = dat_con,
                   control.predictor = list(compute = T),
                   control.compute = list(dic = TRUE, waic = TRUE),
                   verbose = FALSE)
  
  cw_5_loop[[i]] <- data.frame(variable = rownames(m_5_loop$summary.fixed),
                               mean = m_5_loop$summary.fixed$mean,
                               q0.025 = m_5_loop$summary.fixed$`0.025quant`,
                               q0.975 = m_5_loop$summary.fixed$`0.975quant`) |>
    mutate(p_value <- map_df(lapply(m_5_loop$marginals.fixed, 
                                    p_value_function),
                             ~as.data.frame(.))) |>
    rename(p_value = `.`) |>
    filter(grepl("duration", variable) == TRUE) |>
    mutate(duration = as.numeric(str_replace(variable, "duration_5", ""))) |>
    mutate(or = exp(mean)) |>
    mutate(lwr = exp(q0.025)) |>
    mutate(upr = exp(q0.975)) |>
    mutate(bonferroni = ifelse(p_value > duration_alpha_level, "not", "significant")) |>
    mutate(cluster = i)
  
}

cw_5_comparison <- map_df(cw_5_loop, ~as.data.frame(.))

# write.csv(cw_5_comparison, "data/results/cw_5_comparison.csv", row.names = FALSE)
# cw_5_comparison <- read.csv("data/results/cw_5_comparison.csv")

# just a note —
table(cw_5_comparison$bonferroni)
# none significant after bonferroni

ggplot(cw_5_comparison, aes(x = duration, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  ggtitle("Cluster 5 Critical Window Exposure") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Duration of Exposure \n(number of weeks exposed 24-29)") +
  # ylim(0.1, 3) +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1)) +
  facet_wrap(~cluster, scales = "free")


```

Run the model for increased duration of exposure during cluster 10's critical window for all other clusters.

```{r}

## cluster 50's critical window
cw_10_loop <- list()

for(i in 1:12) {
  
  dat_con$duration_10 <- apply(cw_10, 1, function(x) length(which(x==as.character(i))))
  
  dat_con <- dat_con |>
    mutate(duration_10 = as.factor(duration_10))
  
  m_10_loop <- inla(preterm_code ~ duration_10 +
                     poverty_rate_sc + race_ethn, 
                   family = "binomial", data = dat_con,
                   control.predictor = list(compute = T),
                   control.compute = list(dic = TRUE, waic = TRUE),
                   verbose = FALSE)
  
  cw_10_loop[[i]] <- data.frame(variable = rownames(m_10_loop$summary.fixed),
                               mean = m_10_loop$summary.fixed$mean,
                               q0.025 = m_10_loop$summary.fixed$`0.025quant`,
                               q0.975 = m_10_loop$summary.fixed$`0.975quant`) |>
    mutate(p_value <- map_df(lapply(m_10_loop$marginals.fixed, 
                                    p_value_function),
                             ~as.data.frame(.))) |>
    rename(p_value = `.`) |>
    filter(grepl("duration", variable) == TRUE) |>
    mutate(duration = as.numeric(str_replace(variable, "duration_10", ""))) |>
    mutate(or = exp(mean)) |>
    mutate(lwr = exp(q0.025)) |>
    mutate(upr = exp(q0.975)) |>
    # mutate(bonferroni = ifelse(p_value > duration_alpha_level, "not", "significant")) |>
    mutate(cluster = i)
  
}

cw_10_comparison <- map_df(cw_10_loop, ~as.data.frame(.))

# write.csv(cw_10_comparison, "data/results/cw_10_comparison.csv", row.names = FALSE)
# cw_10_comparison <- read.csv("data/results/cw_10_comparison.csv")

# just a note —
# table(cw_10_comparison$bonferroni)
cw_10_comparison$fdr <- p.adjust(cw_10_comparison$p_value, method = "fdr")
table(p.adjust(cw_10_comparison$p_value, method = "fdr") < 0.05)
# one significant after bonferroni — cluster 10, exposure during entire critical window

ggplot(cw_10_comparison, aes(x = duration, y = or)) +
  geom_point() +
  geom_linerange(aes(ymin = lwr, ymax = upr)) +
  geom_hline(yintercept = 1) +
  ggtitle("Cluster 10 Critical Window Exposure") +
  ylab("Odds of \nPreterm Birth \n(≤34 weeks)") +
  xlab("Duration of Exposure \n(number of weeks exposed 9-14)") +
  # ylim(0.1, 3) +
  theme_classic() +
  theme(axis.title.y = element_text(angle = 0, vjust = 0.6, hjust = 1)) +
  facet_wrap(~cluster, scales = "free")


```
